requirements.txt
pytorch
numpy
scikit-learn
scipy
joblib